support for custom validators and inputs
Enables using a custom input in a paper-input-container. Either
provide attr-for-value to observe input value changes for validation
and label floating, or if unset the `input` event is used. See demo
for an example.

Also enables iron-validator-behavior support. The validator takes
precedence over constraints, if provided.

Renamed paper-autogrow-textarea to paper-textarea to be more
consistent with paper-textarea.